added technologies and fixed the nav
