+++ components/Layout.tsx
@@ -1,8 +1,7 @@
 import Head from 'next/head';
 import Link from 'next/link';
 import React from 'react';
-import Navbar from './nav';
-import Hero from './hero';
+import Navbar from './Nav';
 
 const name = 'Mehran Shahidi';
 export const siteTitle = 'Mehran Shahidi';
@@ -16,7 +15,6 @@ export default function Layout({
 }) {
   return (
     <>
-      <Navbar />
       <Head>
         <link rel="icon" type="image/png" href="images/profile.png" />
         <meta
@@ -30,7 +28,8 @@ export default function Layout({
         <meta name="og:title" content={siteTitle} />
         <meta name="twitter:card" content="summary_large_image" />
       </Head>
-      <div className=" h-auto mt-10 dark:bg-gray-800 dark:text-white">
+      <Navbar />
+      <div className="h-auto mt-10 mb-8 dark:bg-gray-800 dark:text-white">
         {children}
       </div>
     </>
